Move data structs behind model module
There were too many structs in the root module and it started to detract
from more common and appropriate structs like Eu4Extractor and Eu4Date
and CountryTag, etc.
