+++ src/eu4date.rs
@@ -12,7 +12,7 @@ pub const EU4_START_DATE: Eu4Date = Eu4Date {
 
 const DAYS_PER_MONTH: [u8; 13] = [0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
 
-/// A date in EU4
+/// Struct specialized to parsing, formatting, and manipulating dates in EU4
 ///
 /// A date in EU4 does not follow any traditional calendar and instead views the
 /// world on simpler terms: that every year should be treated as a non-leap year.
