+++ src/province_id.rs
@@ -1,6 +1,13 @@
 use serde::{Deserialize, Serialize};
 use std::fmt;
 
+/// An province numerical identifier
+///
+/// Handles negative identifiers by taking their absolute value.
+///
+/// ```rust
+/// let _ = eu4save::ProvinceId::new(10);
+/// ```
 #[derive(Debug, Clone, Serialize, Hash, Eq, PartialEq, Default, Deserialize)]
 #[serde(from = "i32")]
 pub struct ProvinceId(i32);
