+++ src/query.rs
@@ -1,7 +1,7 @@
-use crate::{
-    Country, CountryEvent, CountryTag, Eu4Date, Eu4Save, LedgerData, LedgerDatum, Province,
-    ProvinceEvent,
+use crate::models::{
+    Country, CountryEvent, Eu4Save, LedgerData, LedgerDatum, Province, ProvinceEvent,
 };
+use crate::{CountryTag, Eu4Date};
 use once_cell::sync::OnceCell;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
